Show tooltip when dataset name overflows
Fix #626
