+++ mock/datasets/no2.data.mdx
@@ -110,7 +110,7 @@ layers:
         - "#050308"
   - id: no2-monthly-diff
     stacCol: no2-monthly-diff
-    name: No2 (Diff)
+    name: No2 (Diff) - let's make this title reaaaaaaly long
     type: raster
     description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc sodales semper risus, suscipit varius diam facilisis non.'
     zoomExtent:
